Major re-write splitting up the app found in create_quiz.elm into a more well organized set of modules from which to build.  Now the major quiz components of questions, answers, and text modules are separated into:
  /decode.elm
  /encode.elm
  /field.elm
  /model.elm
  /view.elm

files and in the case of the text component also /update.elm.  The
noticeable difference includes far fewer Msgs in create_quiz.elm update
function, which will now be handled by the function defined in
text/update.elm. The rest of the existing view functions will soon be
re-used with some effort.