Update Auth and add Endpoint and Config modules
A handful of changes woven in from a no longer relavant branch.
Endpoint gives us a way to build urls for external API calls.
Config wraps configurations passed in at app initialization.
The Auth module is expanded with a communication layer for passing success and
failure messages up from typescript.
None of this works with the Main yet!