Add frontend roles
Add a Role module built around a type for student and instructor roles.
Update Viewer type to include a role.
Add protectedStudentApplication and protectedInstructorApplication Pages. Any time
a page should only be accessible the first can be used. Any a page is for instructors
only, the latter can be used. The protectedApplication is still available for when
a page can be viewed by any authenticated user.
Update decoders in Api module to pull the role off the JWT. At the moment,
the roles are expected as a "role" claim with a value of "student" or "instructor",
but we can update as needed.
Add example pages that use the new Page variants.