Restructure text reader and start on websockets
- Move TEA parts into module with text reader page. Some parts are still pulled
from a separate Model page, and these might stay there or be brought in.
- Start work on the websocket communication layer. The bburdette/websocket
library would only allow us to pass in a String for content, when we really
wanted to  pass in a Value. The code for this libary was vendored in and modified
a bit. Much work remains around websockets. Their endpoints and communication with
the outside world should be organized into a module like we do with HTTP.
- The UI needs a lot of work here, but this is blocked until a websocket
connection with the backend can be established.
- The changes in TextReader/WebSocket.elm, TextReader/Msg.elm, and
Profile/Flags.elm can mostly be disregarded. These modules won't be used
eventually, but have been kept for reference for now.