Add persistence and centralize profiles
Add persistence for roles and ids. Roles will determine what a user can and
cannot see. For now they are used to fetch the right type of profile.
Request the profile on login and on persistence. Move profiles to Shared where
they can be accessed as needed throughout the application.
Rearrange decoders a bit and move fake profile to Shared. This fake profile
can ultimately be removed when we integrate with the backend. For now, student
and instructor profiles will need an empty profile, but this can probably be
improved handling cases where the profile has been requested from the server
but has not arrived.
Fix the protectedApplication to call the pages load function when a load event
occurs.