Slashed URLs
We've decided to ignore Django's advice to use slashed URLs. In doing so
I removed all of the slashes from URLconf. No changes have been made to
the frontend Elm code. There has been a change made to the Django
settings file `ereadingtool/settings.py` to set `APPEND_SLASH = False`
as it's default value when the CommonMiddleware middleware package is
installed is `True`.