Add show help persistence
Persist whether a user should be shown help or not in localstorage and bring
the flag into Config.
Add a show help checkbox to the profile page to allow students to set their
preference.
Add a show help hint to the profile page explaining the checkbox.
Add save functionality to protected pages.