Dead Code
There appeared to be dead code around the class attribute
`raise_exception` in the classes `StudentAPIView`,
`StudentPerformancePDFView`, and `StudentAPIConsentToResearchView`.
Search and replace via VSCode was unable to find use of this class
attribute. That being the case it was removed in all class definitions.
Another thing to note was its corresponding comment which said:
```
returns permission denied HTTP message rather than redirect to login
```
This may be handled in a different way and will warrant a separate
GitHub issue once the problem arises in testing or practice.