Move navbar to Shared module
Move the navbar to the Shared module. The shared implementation shows an
unauthenticated, student, or instructor navbar depending on the session
and user role.
Remove the placeholder navbar and debugging information that was displayed
below each page.
Remove exposeToken function because it was unused.
Fix wrong title for instructor profile page.
Fix wrong link to instructor sign up page.