JWT authentication
This commit brings in the start of JWT authentication. The new file
directory `auth` contains a single file `ProducerAuthMiddleware` with
the sole purpose of verifying the JWT token passed from the front end.
It creates and ASGI application which wraps the URLRouter and hands it
up to the next thing to wrap it itself (`AllowedHostOriginValidator`).

Initially I didn't think I could use the same `user` object as the
original `AuthMiddlwareStack` could from the previous implementation.
Thankfully, we can easily look up the student with information provided
by the JWT and get all of the necessary fields. Validation of the
expiration is trivial.

The only thing that remains is the case of if the user already exists.
Do we need to go through the whole process again and get the entire
`student` object just to give the green light? Seems obtuse.