jwt_auth in a working position
This commit finalizes the jwt_auth branch. We can now validate a user
based on a JWT in the querystring of their HTTP Upgrade 101 request to
change to a Websocket connection. The final touch was to find what kind
of error messaging we can get from the backend. Turns out an
HttpResponse object was not feasible, intead we `accept()` the
connection, send a message indicating th user is not authenticated, and
then `close()`.