Push instructor vs student auth higher up
The purpose of this commit is to change what was once a double back from
`post_success()` to `post_error()` after a user was discovered to have
accessed the wrong portal. Rather than wait for that `post_success()`,
the solution here is to hit the database and check the `user` object
against the path of the `request` object.