These changes will divide instructors and students
It'll be done earlier on in the application logic for allowing or
denying a student access to the webapp. Students and instructors were
previously allowed through `post_success` and then reverted to a
`post_error` if it turned out they have a particular administrative
attribute. Now the request path is checked against the user's DB entry
to see if the `is_staff` flag is set.

Tests needed to be updated so that they'd work. Specifically, if an
instructor is created by way of `new_user` (as they always are) the
kwarg `is_staff=True` will be passed along.

Along with this change comes the resolution of issue #172, at least
partially. Inheriting from `View` as opposed to `EreadingToolAPIView`
gives us the ability to inherit the `method_decorator(csrf_exempt)`
which will allow prevent CSRF token errors we were previously dealing
with.