APIView brings in more, Student view loses redundancy
So there were two identical PUT methods in the file user/views/api.py
and I brought those up to the APIView class. Unfortunately, I still
needed to use a decorator for the classes that use PUT in the file
api.py The only thing gained here is the removal of duplicate code.
Possibly at the expense of muddling models with the APIView. But I
consider it a win since POST behaves similarly.