Weird changes
As with all things modern, I had to do some weird stuff. For instance, I
needed to change the command and entrypoint because docker-gen was
crashing. https://github.com/jwilder/docker-gen/issues/201

I also am now trying to mount the conf file as a volume to get it copied
over since the COPY command in the Dockerfile.nginx seems to be ignored
by docker-compose. Starting the container on it's own will lead to the
file being copied. It does not appear when started with the other
containers (via docker-compose).

The file Dockerfile.nginx will no longer be necessary.