Redis on the backend
This commit changes things regarding the relationship between the Django
backend and the Redis container.

One crucial thing to note is the nginx-proxy container absolutely needs
to specify VIRTUAL_PORT if there is more than one exposed port on the
container. This was the source of much confusion, even though it's
expressed clearly in the documentation. Rats.

Another thing this commit adds is the appropriate host ('redis') which
is named after the image's service name in the `docker-compose.yml`
file.