CSS and correct string
Changed the strings so that they accurately reflect the page that is
being looked at by the user. Also updated the CSS. Upon inspection I
found that the login title was neither positioned or a CSS flexbox/grid
elemnt. That being the case I made it a grid and centered it
accordingly.