PDF access verified by JWT against DB
These changes converted the `jwt_validation` function to a synchronous
version and placed it in the same `auth/` directory in `normal_auth.py`.
This code does the same validation of the time included in the JWT,
verification of the secret_key against the jwt's, and we extract the
user's `pk`.

Once the `pk` is extracted from the JWT it is compared against the one
passed from the frontend. If they're not the same then the JWT was made
for a different profile and will be rejected accordingly.