Dependencies, more clear use of HttpResponse
Instead of just using HttpResponse, we're using HttpResponseForbidden.
This makes it more explicit that the user is not permitted to access the
webpage, rather than just returning a status of 403. They're effectively
the same, but the code is cleaner. Also, cleaned up up some unnecessary
dependencies in this commit. For one example, we no longer needed the
LoginMixin as we're authenticating with the JWT.