Adding dependencies to dockerfile
Without this dependency the container will error out when attempting to
get a PDF version of the performance report. Instructions are to add the
executable to your PATH, but it appears to work with just an `apt
install wktohtmlpdf`. Ran this command on the development container and
it seems to work.