Not wrapt
Well, tried to add the meta data with `@wraps` from `functools` but that
barfed with some error mentioned in the previous commit message. Now I
tried playing around with `wrapt` but for some reason it doesn't want to
be installed. `pip install wrapt` confirms that it was installed
successfully but the environment does not have the package. Looked
around for a little while and gave up.

This wrapt stuff is only helpful for debugging anyways, it copies over
metadata from the contained function. Seeing as it's just us, this is
not a path I care to continue.