JWT decorator added
It exists now, but I'm still debating if we should check the database to
see if the `pk` matches that of the JWT. I'll need to find best practice
for that. Another thing is that we're only giving it a single status
code, support for many would require a data structure.

Currently checks for a JWT in the request's querystring. Not sure how to
wrap this around every API endpoint unless the JWT is always passed.