Unverified, but likely working JWT auth decorator factory
Major changes to the auth mechanism were made in the commit. Now we
check the Authorization header field of the request or the querystring
depending on which is provided.

**NOTE** we assume that the only querystrings hitting this backend will
be tokens. Further modifications will need to be made if that is to
change.

There is also a verification of the user's JWT against the `pk` that
they're requesting. The JWT contains a `user_id` which is used to access
the whole user's profile. Their status as student or instructor is
determined in the function; that info is cross referenced with the `pk`
passed in the request. Note that this results in a hit to the database,
so it could be a potential performance bottleneck. But I think this is a
worthwhile trade off, as simply incrementing the pk in the request with
a valid JWT will yield access to anyone's profile. Open to suggestions
on this.