PDF endpoint modifications
Remembered that we have to check the PDF endpoint differently. We can't
use the decorator factory because we can't reply with just JSON. There
is a need to serve up the appropriate HTML webpage. Because of this,
there is now a local function in `student_performance.py` which does JWT
validation.