Localize datetimes
Add app init to determine user's timezone and store it in Shared.
Update datetime formatter to take a Zone and a POSIX time. Update all
calls to the formatter to take the zone to convert it correctly.
Update decoders to decode to Posix instead of DateTime.
Remove DateTime library since we aren't using it anymore.
Add tests to validate new datetime formatter implementation and add
localization tests.