New SendGrid token validation
Previously we'd get a link from SendGrid, open that link and fire off a
request to the backend including the uidb and token in the GET request.
Then the server would reply with a redirect to scrub the token from the
URL. The purpose of this is to remove the token so any additional pages
being loaded don't leak it in the `referer` field.

We've done away with the redirect and simply reply with a JSON error.
Unfortunately we've reopened a security hole since we allow for the
referer to leak the security token.

We also no longer redirect to a compiled Elm page if things went wrong.

Solutions to the referer bug:
* Upgrade to Django 3.0 and set SecurityMidddleware setting https://docs.djangoproject.com/en/3.1/ref/settings/#std:setting-SECURE_REFERRER_POLICY
* Nginx (maybe?)
  https://serverfault.com/questions/869463/nginx-rewrite-http-referer
* Go back to the exit page
  https://geekthis.net/post/hide-http-referer-headers/#exit-page-redirect