Domain independent emails now possible
We had previously been depending on
django.template.context_processors.request to determine the domain in
`ereadingtool/user/templates/registration/password_reset_email.html`.
Because the VIRTUAL_HOST environment variable is the domain we'd like to
use (without the prefix 'api.'), we set it accordingly. The alternative
to using VIRTUAL_HOST would be to rely on the previous `domain` setting,
but I'm unsure how it will behave behind the reverse-proxy.