Far cleaner solution
Rather than doing a hacky hardcode of the URL, just save the
`FRONTEND_HOST` in an environment variable. Be sure to set this
environment variable in your local env, else `domain_override` will be
None. This will lead to an incorrect port number, since it'll query the
host running the Django app and receive `localhost:8000` when we really
need `localhost:1234`.