DELETE header added to cors configuration
There's a file at the top level called cors.conf It's copy pasted onto
the reverse proxy via `volumes` in the docker-compose file. Adding
DELETE to this list should allow us to run DELETE http requests in the
few places in which they're done. Namely, deleting an entire text and
deleting the lock on editing a file.

Not sure how to test this PR on localhost. We won't get any CORS
conflicts until attempting to access resources from a subdomain.