WSGIRequest error
The reason all of our test are failing is because they're wrapped by a
decorator that made too many expectations. It was previously assumed
that you could always get `scope` and `META` out of the request object.
As it turns out, when the test is run it uses a WSGIRequest (which
actually inherits from HttpRequest). Suffice to say, the previously
listed fields are not members.

Now we just skip the HttpResponse usually given when something goes
south in the decorator factory. Please do note that we still assume
HttpRequest to have those members. If in fact that is not the case there
is much trouble involved.

TODO: test around this. Aren't ASGI objects always used instead? Is this
due to middleware?