Password reset email works
Needed to remove some trailing slashes to get the `/login/student` URL
to go to the right place. Then added a `post_error(..)` field to bail
out of form validation if the token provided by the frontend is invalid.
Made sure to add an error to the form appropriately, but that caused
api.py to barf. Wrapped that all in a `try` (as it should have been),
but the issue didn't persist once things were written correctly. Doesn't
hurt.