Testing...
Removed what I thought was good user protection in the auth system we've
been creating. Looks like we'll go with minimal changes and try to break
it once it's running, rather than preventative. Here I'm referring to
checking the user's ID in the DB against the ID that was provided in the
JWT. Also, not sure if this matters (as it doesn't work on my machine
anyways) but there is a link to a very cryptic issue regarding SQLite3
and testing.

https://github.com/weluse/django-nose-selenium/issues/8

This only has to do with selenium testing, and it doesn't seem to have
made a difference, but worth noting. Our tests may or may not work well
with the generated in-memory database created by Django. So we can use
`TEST_NAME` in settings.py to force more stable behavior apparently?