Producer Auth error
The ideal solution to finding an invalid JWT during a websocket
connection is to discontinue the connection and return JSON to the
client so they can make a sensible decision. We're part way there,
previously we just disconnected and that was all. Now we intentionally
return None, which seems to result in a 403 Access Denied.

Much more work needs to be done (some sort of global indicator that the
token is bad) to establish the connection, send a message, then
disconnect.