Lemmas
We got em! Took forever, couldn't figure out why the thing was upset
about a uniqueness error. I needed the lemma to exist, the addition to
unique_together needed to have it so it wouldn't barf, either way it was
upset. Commented out the previous migration altering the uniqueness and
we were good to go.

https://stackoverflow.com/questions/53258853/problem-with-updating-unique-together-constraint-django/64813956#64813956