Virtual table report_student_performance
After many hours trying to track down this bug, I still haven't squashed
it. From what I can tell, it just cannot find the virtual table
report_student_performance due to some underlying change. I have a very
strong reason to believe this is related to migrations in some way,
because deleting `report/migrations/000*` allowed us to run a majority of
the tests. Keeping `0001_initial.py` and
`0002_student_performance_report.py` yielded almost the same error but
ran none of the tests.

All I've managed to discover is that technically the table doesn't
exist because it's a view. Why this is accepted on other branches is
beyond me, perhaps there is a versioning error I'm not seeing.

Things I've tried to remedy this:

* Dropping the django* tables
* Removing the view entirely
* Reseting the migrations as described in the link below.

Nothing has worked, so for now the tests are `@skip`ped.

https://simpleisbetterthancomplex.com/tutorial/2016/07/26/how-to-reset-migrations.html