Websocket disconnect response
This patch fixes what would have otherwise been an infinite loop if the
user doesn't provide a valid JWT. We used to return a string and that
would lead to an `ValueError` returning None to the wrapping
Middleware. Well, turns out this dropped us in an infinite loop instead
of closing the 101 upgrade. Now, we still allow for that infinite loop
but should encounter it. Perhaps a network error results in that path
being taken, in which case you've got bigger problems anyways.

Now, we open the connection pass JSON indicating that the JWT is
invalid, and promptly close the connection. We're not able to return an
`HttpResponse` object with a status of 403 from what I can tell.