JWT validation
Made further changes that include limiting the decorator in all places
to just be a 403 with the same error message. Similar to the previous
commit but now applied everywhere. Did away with the
`LoginRequiredMixin` so now we aren't redirected to the login page when
Django thinks the session is expired.