New fix
Now we preserve the number of sections read until they've clicked start
in the intro state. One thing to note: If they enter the intro state
their timer for Last Read will be set to the that moment. Their history
of questions correct will also be wiped out.

We could preserve Last Read and Questions Correct until Sections
Complete is also rolled over, but we would need to reimplement a couple
of functions.

Submitting a PR on this as it's bare minimum and can be rejected if we
decide this is insufficient.