Lemma in grammeme
This fix moves lemmas into grammemes. Thought this was going to ruin
migrations, but quickly realized that TextPhraseGrammemes is inherited
by TextPhrase, so the model underneath doesn't need to change. For some
(concerning) reason the uniqueness constraint is now failing unless the
`lemma` is included. This should not be the case as the three other
fields should be sufficient. Continually worried about out dataset.