Change read again to a link and take id from page
The ID was taken from the text and we may have had a race where it was the 0
from the empty text. We have the id on the page, and we can rely on that a to
be there. The onclick to route was also changed to a link per #285.