Authentication on the admin panel
There's NONE! Mwuahhaah. Bit of a challenge to figre out. The _only_
place I've seen this on the web is at the following SO post. My thoughts
are "if we remove the normal auth then we can use our own". Currently
think we can allow or deny based on if they are a superuser in the
readeruser class. Remains to be seen!