Add close all websockets command
Add a command to disconnect all open websockets. We need to send a disconnect
when a reader finishes a text, and this can happen on any page navigation. We
don't have a teardown mechanism in elm-spa, so for now we clean up when we init
the new page -- not ideal, but it will work for now.