Failed attempts at accessing view data
Keeping these around as a reminder. Views are different -- what needs to
happen is a custom SQL stmt to create a view. Then write the model that
is unmanaged around that view. This is alternative to using the data
provided by the tables that have already been created. You can see this
play out with the foreign key on text_reading having yet to be
evaluated.