Fix valid word indexing bug
We needed to treat instance and occurrence as different concepts for valid
words when we reindex words after applying compounds. The reason is that
an instance is used to look a translation and needs to be preserved. An
occurrence indicates position in the text and needs to be updated. When valid
words become part of compound words, the instance and occurence can drift apart.
Added comments here as well to describe how this all works.