Pinned pyJWT and removed --upgrade
pyJWT moved to 2.0.0 which resulted in breaking changes. The `--upgrade`
flag in CircleCI/config was trampling my request to keep this older
version. Also moved the CI tests to use 3.9

still fails?!

updating circleci to use python3.9.1

missed a config

don't upgrade

For some reason we upgrade everything in the build environment. This was
sending a pinned 1.7.1 up to 2.0.0 where breaking changes exist

tough upgrade decision

So pip was ignoring my dependency pin at 1.7.1 due to the `--upgrade`
flag. Did a little research and found that this is not the easiest
problem. As it turns out there is an entirely new dependency resolver
that was released about a month ago! I'm unsure how this would effect
our project so I'm erring on the side of caution. That being said I
ditched the `--upgrade` because the versions should be recent as they're
freshly installed in the build.

need vision

surprisingly can't see into the dev environment, so adding print
statements

undecode

forgot that I don't need to decode it because we SHOULD be on 1.7.1