Report problem on invalid words and add "/" case
Update parser to fail on invalid words with a problem. We want to keep invalid
words to show them, but we don't want to tag them. We already have the word
before we parse, and we can just default to this text including any leading or
trailing punctuation. This simplifies the number of cases we need to consider
for invalid words.
Add a case for "/" in the middle of a word. This can happen in ratios.