Texts started and complete
The `StudentPerformanceReport` now returns the texts a student has
completed in various timeframes, along with the texts that they have
started. Note that texts started are counted within some time period,
and they don't carry over to a more recent one.

Note that this has breaking changes. The data returned to the frontend
is now missing, there are different fields / missing fields.
Consequently the student profile does not load. To see changes in the
PDF you can attempt to download the performance report. If you're logged
in you should have a valid JWT, but you'll need to change the student id
from 0 in the URL to your student ID.