Post error update
When the form is found to be invalid because an entry already exists in
the table, we now use the child class's `post_error(..)` method. Prior
to this we used the `post_error(..)` method defined in
`user/views/api.py`. Had to change the type signature, but based on
search for `.post_error` the only other place it's used is in
`forgot_password.py`. Changed the function signature there but it just
ignores the extra two params and sends JSON anyways. The reason for all
of this is we needed the request to check the token since the new
`post_error(..)` in `user/views/instructor.py` is JWT auth protected.