Restyled Page titles
Needed to restructure some Elm code, but it ultimately worked out.
I know I should be hesitant to change fonts, but it was quite easy to
make the Signup page fonts match the ones on the login pages. Now the
titles reside within their respective boxes (`signup_box`, or
`login-box`). This forces the spacing from the navbar to be the same
across these pages.