Title margins
This shores up the text below the signup pages and acknowledgement pages
so that they're equivalent to their respective titles. Also put the
acknowledgement text in a body like the signup pages. This made the text
as narrow as text on other pages. Gives the benefit of setting a padding
on that whole container.