Endpoint added
We've always had this endpoint. Oddly, it seems to delete some of the
entries, in various tables -- but not all.

For example, calling this endpoint with a text group to delete will
remove a row from `text_textwordgroup` as well as `text_textgroupword`.
But that does not remove the individual entries from `text_textphrase`.
It will remove the merged word (a single db row), but it will not remove
the two entries that were added just prior to the merged word.

Attempting to remove `text_textphrase`s by hand results in the foreign
key error.