Using views to solve filter bug
Technically it means nothing to have a primary key on a View in the
context of SQL. But to Django's ORM we can attach a primary key even if
SQLite3 has understanding of it. Again, this came from the holy grail of
Django DB View [blog
posts](https://resources.rescale.com/using-database-views-in-django-orm/)

Steps to do this:
1) In DB Browser for SQLite right click the view to update
2) Modify View => Ok
3) In the Execute SQL add the following as the first entry after `SELECT`:
  ```
  row_number() OVER() as id
  ```
4) Run the query

This will basically give a unique value to every row in the view, which
Django can use as a primary key and facilitates getting objects from the
view.

Now we use these views to get texts and return them back to the existing
machinery.