Translation service processed flag on frontend
At the text level we now have a flag indicating whether or not a text
has been translated. To test this, check a text that has already been
translated. Then add a text section. Check to see the flag is false.
Then access the database and find out the id of the text section just
added. Use the Django command --text_section to explicitly run the
translation service on the section. Depending on the number of words
added, it may take some time.

Here's an example:

```
python manage.py create_text_phrases --text_section <ID>
```

The flag should now be set to true.