Response to frontend from unmerge
Not the cleanest implementation- ended up deleting many unused fields to
match the JSON required by the frontend. But on the whole it should
work.

To get here the route must have an int on the end. We use this value to
get info about the grouped words members. We assume that the delete
process will work.