Responding to frontend from unmerge request
Indicates that a groupd word was either succesfully removed from the
database or not. We know that the end of the path must contain a number,
so we use that to look up all of the associated text words.

Replies with a `True` or `False` indicating that the grouped word was
removed or not. We also return the corresponding error message in the
"error" field.