Flashcards pdf
We can now give students a summary of their flashcards if they visit
`host:8000/flashcards.pdf?token=JWT`. Similar technique to the
performance report, we generate a PDF from templated HTML.

A new requirement came in on this one, `PyYAML`. We're using it to get
the frontend host from the `docker-compose.yml` file.

rebasing... -.-