Hide flashcard gloss option when no translations
We may have cases where a translation is not shown to a student when they gloss
a word. This happens when no translation has been selected as correct for
context. We want to hide flashcards in these cases so that a student doesn't
get flashcards without a translation.
A few styles were added to flashcard options here as well/