CSV flashcard file
This commit brings in flashcards as a CSV filetype, even though the
delimiter is actually a `|` character. Reason being we have commas
littered throughout the text bodies. Quizlet allows us to set a custom
delimiter anyways, this seemed like the most reasonable and least buggy
option.

This was being troublesome before because ContentFile was calculating an
incorrect size. Cyrillic unicode character are two bytes and it was only
counting one. We transform the string to bytes and take that size on the
backend instead. We also drop all `"` characters because `StringIO`'s
`getvalue()` function was wrapping everything in quotes. `"` don't show
up in the text but if they do they're dropped sorry about that. This
makes the flashcards look much cleaner after importing.