Adding `admin_panel/`
This pretty much nests the app with a different auth mechanism. It's
challenging (couldn't see how to do it) to have sessions alongside JWTs
since the middleware wants to do only one thing. Instead we've got the
admin panel hosted on a different port interacting with the same
database. It's really more of a stop-gap.