Squashing a bunch of commits
They probably should have been removed with a fixup

Change db location

it's up a dir

Testing cert on

Looks like we're making too much noise

on second thought, do away with third A record

Adding to ALLOWED_HOST for second dev server

admin-panel container

now added. need to register subdomain

Admin panel up with styling

Found out that the styling from the admin panel needs to be served from
somewhere when DEBUG != True. At this time there are so few users using
the admin panel that the development server can handle this, so we're
going to move those files over to the admin panel's `static` dir. In the
future they should be probably be bundled and sent with the other static
files by the frontend container.

Messing with allowed hosts to permit admin and api

Admin panel styling

This was actually very painful. For some reason Django just didn't want
to serve out of the project's static directory. Had to inherit and
extend the admin panel's template for it to find the css. Tried about
half a dozen other implementations, this is the only one that worked.

https://stackoverflow.com/questions/7357057/overriding-admin-css-in-django

admin URL is only subdomain

Wrong name for inherited template

It's the order that matters

two settings.py ...

whitenoise

attempting to use this fancy library to serve static assets

wrong settings.py facepalm