Admin panel up with styling
Found out that the styling from the admin panel needs to be served from
somewhere when DEBUG != True. At this time there are so few users using
the admin panel that the development server can handle this, so we're
going to move those files over to the admin panel's `static` dir. In the
future they should be probably be bundled and sent with the other static
files by the frontend container.