DELETED ALL STATIC FILES
Pehaps this should be on its own branch, but we're expirimenting with a
second dev server at the moment. Trying to reconfigure the static files
being shared by django so that we can get the admin panel working.

Currently everything is brought the users with Parcel. We still need to
serve CSS out of the admin panel's container, preferably by way of
WhiteNoise. We'll see how it goes.