Add difficulty info to sign up page
We want to give students an easy way to evaluate their preferred difficulty
when signing up. This commit gives the an info toggle and difficulty info
for the selected difficulty.
Also changed the width of the welcome text above the form to align with the
inputs and fixed a bug where the show password eyeball fell out of line
on mobile.