Student Profile tests at milestone
Mostly covered the entire student profile page. This includes navigating
the hints, checking links, presence of various items. Dynamic content
is the focus at the moment, whether or not links direct to the correct
location is the type of thing that may fail logically. The orientation
of features in the app, color, width and the like is less thoroughly
examined.