Student side covered, added files for instructor
These tests mostly cover the student side. Of course more could done to
verify styling, but the functionality has largely been covered. Two
caveats- the demo text needs to be completed by the cypress user each
time it is touched.

** Important **
The other caveat is that it'll make a metric ton of
entries in the database if these tests are frequently ran. May be worth
scrubing the user and starting over once this is in prod and updates are
minimal.