Tag error fixed
An index error was being thrown because we weren't able to filter them
on the database view. This commit deletes them from the filter and takes
the set of texts that match other constraints (like difficulty and read
status). Following that we take the intersection of that set and one
which includes all the texts for that difficulty but have been filtered
by tag. This gives us two sets of texts with the same difficulty, but
one has been filtered on tags and the other on read status. Again, we
check for the overlap.