Opacity bugfix
Somewhere in our production build there is use of cssnano, which has a
bug that reduces opacity to 1% from whatever other setting. The fix to
this is either to use cssnano@nightly, or simply change the opacity to a
decimal rather than percentage. So it was converted to a decimal and the
problem has been averted.

https://github.com/NMFR/optimize-css-assets-webpack-plugin/issues/118