Hotpatch content-editor search
This hotpatch fixes the 500 on the content-editor search page.
The difficulty key was being used to filter items on the backend, but
that doesn't exist for instructors. Students have the filter set to
their preferred difficulty, instructors have no such setting. The fix is
to simply see if the key exists, if not proceed through the method.