Cypress tests
This commit brings in over 50 Cypress integration tests, supporting
documentation along side a visual regression plugin. Using these tests
we can confirm frontend functionality, assure that modifications to CSS
are not damaging to the site's appearance, and keep track of their
differences with a report. Fitting the tests in CI proved to be
challenging since the database would need to be built or on the runner.