Tests managed temporarily
Django builds up a temporary database for testing based on code found in
migrations files. Unmanaged models, like those used with database views,
are not created by the tests. In order to do so, we needed to specify a
custom test runner which would iterate through all of the models in our
various apps setting their meta data `managed=True`.
https://gist.github.com/raprasad/f292f94657728de45d1614a741928308
https://resources.rescale.com/using-database-views-in-django-orm/