Styling Student Guide
The tabs proved to be a bit challenging. They're now limited in size,
and when the screen shrinks sufficiently small (< 460px) the font
becomes smaller. This was required to get the grid containers to play
nice on the same row when the screen size is less than 340px in width.