gzipped JSON on the text page
The text page is first inline for being gzipped since the content
returned by the server is lengthy. The delay is shown on the student
profile as well as the instructor's page: most notably the "Edit" tab
where the list includes all texts. The speed performance was almost
imperceptible, but there was almost a 90% reduction in file size.
Content-Length uncompressed: 102658
Content-Length compressed: 12414