Extra whitespace at top in text reader
This is caused by the `exception` one receives when attempting to
proceed through a text without selecting an answer. In FF this area has
0 height until the exception actually occurs, similar behavior exists in
Chrome. In Safari the area is noticeable, so it needs to be reduced.

Best thing I could think of: `auto` for the row rather than 10%. Not
sure why FF was showing height of 0, Chrome had a small height... Safari
had an eyesore of a height on mobile.

Padding to navigation at bottom of text reader

The previous fix for the whitespace at the top of the reader bumped the
navigation buttons to the bottom of the screen. Giving them some
padding in this commit.

Frustratingly enough their size and position changes slightly between
sections for some currently unknown reason.

More padding on the bottom

For some reason I need to click next twice on the text reader, I think
it's still too close to the bottom of the screen since previous doesn't
have this issue.

Removing min/max

Perhaps having fixed size buttons rather than using mix/max will help
the navigation on mobile